Refactor menu items for storage, sql editor and table editor
