+++ studio/components/ui/ProductMenu/ProductMenuItem.tsx
@@ -1,13 +1,12 @@
 import Link from 'next/link'
-import { FC, ReactNode } from 'react'
-import { Badge, Button, Menu } from 'ui'
+import { ReactNode } from 'react'
+import { Button, Menu } from 'ui'
 
-interface Props {
+interface ProductMenuItemProps {
   name: string | ReactNode
   isActive: boolean
   isExternal?: boolean
   icon?: ReactNode
-  action?: ReactNode
   url?: string
   target?: '_blank' | '_self'
   onClick?: () => void
@@ -16,45 +15,18 @@ interface Props {
   label?: string
 }
 
-const Label = ({ label }: { label: string }) => {
-  const color =
-    label.toLowerCase() === 'new' ? 'text-brand-900 bg-brand-500' : 'text-amber-900 bg-amber-500'
-
-  return (
-    <span
-      className={[
-        `text-mono 
-        h-[18px],
-        flex
-        items-center
-        justify-center
-        px-1.5
-        rounded-full 
-        tracking-widest
-        text-[10px] 
-        font-medium
-        truncate`,
-        color,
-      ].join(' ')}
-    >
-      {label}
-    </span>
-  )
-}
-
-const ProductMenuItem: FC<Props> = ({
+const ProductMenuItem = ({
   name = '',
   isActive,
   isExternal,
   icon,
-  action,
   url = '',
   target = '_self',
   onClick,
   textClassName = '',
   hoverText = '',
   label,
-}) => {
+}: ProductMenuItemProps) => {
   const menuItem = (
     <Menu.Item icon={icon} rounded active={isActive} onClick={onClick}>
       <div className="flex w-full items-center justify-between gap-1">
@@ -62,12 +34,11 @@ const ProductMenuItem: FC<Props> = ({
           title={hoverText ? hoverText : typeof name === 'string' ? name : ''}
           className={'flex items-center gap-2 truncate w-full ' + textClassName}
         >
-          <span className="truncate">{name}{' '}</span>
+          <span className="truncate">{name} </span>
           {label !== undefined && (
             <span className="text-orange-800 text-xs font-normal truncate">{label}</span>
           )}
         </div>
-        {action}
       </div>
     </Menu.Item>
   )
